backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
numpy
pandas
scikit-learn
httpx
openai
groq
pydantic
python-dateutil
